2_vision/deepfake_detection/requirements.txt
pandas==1.4.4
opencv-python==4.6.0.66
pydot==1.4.2
flake8==5.0.4
PyYAML==6.0